feat: 코드리뷰 반영(문자열계산기)
문자열을 검증하는 부분과 계산하는 곳을 분리
검증시 require 사용
find 함수 사용
"1 + 1 + 1 +" 의 경우
테스트 케이스 추가
