+++ src/main/kotlin/study/step2/CalculatorMain.kt
@@ -4,7 +4,9 @@ fun main() {
     print("계산식을 입력해주세요 : ")
     val value: String = readLine()!!
 
+    Expression.valid(value)
     val calculator = Calculator()
-    val result = calculator.calc(value)
+    val result = calculator.calc(Expression.parse(value))
+
     print("계산식의 결과는  $result 입니다")
 }
